+++ R/utils.R
@@ -22,8 +22,12 @@ disable_non_interactive_quit <- function() {
 #' @return the exact length of the time span between start time and end time
 #' @export
 #' @examples
-#' \dontrun{get_job_duration(get_script_run_time(),get_current_time()}
-#'
+#' \dontrun{
+#' get_job_duration(
+#'   start_time = get_script_run_time(),
+#'   end_time = get_current_time()
+#' )
+#' }
 get_job_duration <- function(start_time, end_time) {
   return(lubridate::time_length(end_time - start_time))
 }
@@ -43,6 +47,8 @@ get_package_scope_var  <- function(key) {
 
 #' Initialize all etl dependencies
 #'
+#' @param project_name name passed to \code{\link{set_project_name}}
+#' @param project_instance name passed to \code{\link{set_project_instance}}
 #' @param script_name name passed to \code{\link{set_script_name}}
 #' @param fake_runtime An optional asserted script run time passed to \code{\link{set_script_run_time}}, defaults to the time this function is called
 #' @param log_db_drv, an object that inherits from DBIDriver (e.g. RMariaDB::MariaDB()), or an existing DBIConnection object (in order to clone an existing connection).
@@ -53,7 +59,9 @@ get_package_scope_var  <- function(key) {
 #'   init_etl("name_of_file")
 #' }
 #'
-init_etl <- function(script_name = "", fake_runtime = NULL, log_db_drv = RMariaDB::MariaDB()) {
+init_etl <- function(script_name = "", project_name = "", project_instance = "", fake_runtime = NULL, log_db_drv = RMariaDB::MariaDB()) {
+  set_project_name(project_name)
+  set_project_instance(project_instance)
   set_script_name(script_name)
   if (!is.null(fake_runtime)) {
     set_script_run_time(fake_runtime)