+++ tests/testthat/helper.R
@@ -32,3 +32,8 @@ if (is.null(salt)) {
   set_package_scope_var("salt", paste0(runif(1), runif(1), runif(1)))
   salt <- get_package_scope_var("salt")
 }
+
+# write a dataframe, referenced by 'table_name' to tests/testthat/directory_under_test_path
+write_rds_to_test_dir <- function(table_name, directory_under_test_path) {
+  get(table_name) |> saveRDS(testthat::test_path(directory_under_test_path, paste0(table_name, ".rds")))
+}